ISSUE-177 Add helm hook to force helm to wait until the install is complete
Previously, running the install or upgrade action would exit helm as soon as all
resources were deployed, not when the job was completed. Adding these hooks causes
helm to wait until the resources are fully deployed. Note that we can only include
install and upgrade (and not any deploy-related action) because otherwise we have a
cyclical dependency (post-install hook + waiting for install to complete).