added .custom.sidecarContainers
added .custom.sidecarContainers to be able to have sidecar containers running together with the pega-web-tomcat container. This is helpful if you want to extract pega-logfiles to a shared volume, so you can send them to a 3rd party (kibana). A second implementation could be running a Stonebranch agent so you have file transfers on a  shared volume.